requirements.txt
Django==4.2.10
django-cors-headers==4.3.1
django_crum==0.7.9
djangorestframework==3.14.0
gunicorn==21.2.0
Hypercorn==0.15.0
requests==2.31.0
uvicorn==0.24.0.post1
